Bug fix in file name edit line.
When a file was dragged/dropped into the edit box for input
file, output file or subtitle file, the application was not
properly notified of the change and acted as if the edit box
was not changed.

As a side effect, if the subtitle file was specified this way,
it was ignored and no subtitle was inserted.