Bug fix on Windows when using HTTP and OpenSSL.
When looking for a new version, the release notes for the new version
are fetched from SourceForge. Due to a new policy, SourceForge now
returns files using HTTPS, not HTTP. The way QtlMovie was linked on
Windows prevented the use of OpenSSL and HTTPS. This is fixed.

Help files were missing from Windows standalone packages. This is fixed.